requirements.txt
absl-py==2.3.1
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
anyio==4.11.0
black==25.11.0
blessed==1.21.0
Brotli @ file:///home/conda/feedstock_root/build_artifacts/brotli-split_1749229842835/work
cairocffi==1.7.1
CairoSVG==2.8.2
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1754231422783/work/certifi
cffi @ file:///home/conda/feedstock_root/build_artifacts/cffi_1725560520483/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1754767332901/work
click==8.3.0
contourpy==1.3.2
cssselect2==0.8.0
cycler==0.12.1
defusedxml==0.7.1
exceptiongroup==1.3.0
filelock @ file:///home/conda/feedstock_root/build_artifacts/filelock_1755216263872/work
fonttools==4.59.2
fsspec==2025.10.0
gmpy2 @ file:///home/conda/feedstock_root/build_artifacts/gmpy2_1745509363766/work
gpustat==1.1.1
grpcio==1.74.0
h11==0.16.0
h2 @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_h2_1756364871/work
hf-xet==1.2.0
hpack @ file:///home/conda/feedstock_root/build_artifacts/hpack_1737618293087/work
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==1.0.1
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1737618333194/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1733211830134/work
imagecorruptions==1.1.2
ImageIO==2.37.2
imgaug==0.4.0
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1741263328855/work
kiwisolver==1.4.9
lazy_loader==0.4
Markdown==3.8.2
MarkupSafe @ file:///home/conda/feedstock_root/build_artifacts/markupsafe_1733219680183/work
matplotlib==3.10.5
mpmath @ file:///home/conda/feedstock_root/build_artifacts/mpmath_1733302684489/work
mypy_extensions==1.1.0
networkx @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_networkx_1731521053/work
numpy==1.26.4
nvidia-ml-py==13.580.65
opencv-python==4.10.0.84
opencv-python-headless==4.11.0.86
packaging==25.0
pandas==2.3.2
pathspec==0.12.1
pillow @ file:///home/conda/feedstock_root/build_artifacts/pillow_1751481981736/work
platformdirs==4.5.0
protobuf==6.32.0
psutil==7.0.0
pycparser @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pycparser_1733195786/work
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1733217236728/work
python-dateutil==2.9.0.post0
pytokens==0.3.0
pytz==2025.2
PyYAML @ file:///home/conda/feedstock_root/build_artifacts/pyyaml_1737454647378/work
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1755614211359/work
safetensors==0.6.2
scikit-image==0.25.2
scipy==1.15.3
seaborn==0.13.2
shapely==2.1.2
shellingham==1.5.4
simsimd==6.5.1
six==1.17.0
sniffio==1.3.1
stringzilla==3.12.6
sympy @ file:///home/conda/feedstock_root/build_artifacts/sympy_1745946051654/work
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tifffile==2025.5.10
timm==1.0.21
tinycss2==1.4.0
tomli==2.3.0
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tqdm==4.67.1
triton==2.2.0
typer-slim==0.20.0
typing-inspection==0.4.1
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1756220668/work
tzdata==2025.2
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1750271362675/work
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.1.3
zstandard==0.23.0
